refactor(#96): reorganize project structure for being more scalable and readable
