+++ cosinnus_frontend/.eslintrc.json
@@ -7,9 +7,23 @@
   "rules": {
     "semi": ["error", "never"],
     "@typescript-eslint/semi": ["error", "never"],
-    "@typescript-eslint/no-unused-vars": ["error"],
     "@typescript-eslint/space-before-blocks": "off",
-    "react-hooks/exhaustive-deps": "warn",
+    "@typescript-eslint/naming-convention": [
+      "error",
+      {
+        "format": ["strictCamelCase", "PascalCase", "UPPER_CASE"],
+        "selector": ["variable"],
+        "leadingUnderscore": "allow"
+      }
+    ],
+    "@typescript-eslint/no-unused-vars": [
+      "error",
+      {
+        "varsIgnorePattern": "^_",
+        "destructuredArrayIgnorePattern": "^_"
+      }
+    ],
+    "react-hooks/exhaustive-deps": "error",
     "no-multiple-empty-lines": [
       "error",
       {
