+++ cosinnus_frontend/src/components/shared/boxes.tsx
@@ -1,8 +1,22 @@
-import React from "react"
-import { Box, useStyleConfig } from '@chakra-ui/react'
+import React, { FC } from 'react'
+import { Box, useStyleConfig, BoxProps } from '@chakra-ui/react'
 
-export function StyledBox(props: any) {
-    const { variant, ...rest } = props
-    const styles = useStyleConfig('Box', { variant })
-    return <Box __css={styles} {...rest} />
+
+interface StyledBoxProps extends BoxProps{
+  variant: any
 }
+
+
+const StyledBox: FC<StyledBoxProps> = (props) => {
+  const { variant, ...boxProps } = props
+
+  const styles = useStyleConfig('Box', { variant })
+
+  return (
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    <Box __css={styles} {...boxProps} />
+  )
+}
+
+
+export default StyledBox
