+++ cosinnus_frontend/src/pages/password-reset/index.tsx
@@ -1,10 +1,6 @@
 import React, { FC } from 'react'
 import { useSelector } from 'react-redux'
-
-import { Link as RouterLink } from 'react-router-dom'
-
 import {
-  Box,
   Heading,
   Button,
   VStack,
@@ -12,23 +8,21 @@ import {
   FormLabel,
   Center,
   Text,
-  Link,
 } from '@chakra-ui/react'
-
 import {
   Formik,
   Form,
 } from 'formik'
 
 import { FormattedMessage } from 'react-intl'
-import { InputField } from '../../shared/input'
-import { TwoColumnPage } from '../../shared/pages'
-import { StyledBox } from '../../shared/boxes'
+import InputField from '../../components/shared/input'
+import SplitColumns from '../../layouts/SplitColumns'
+import StyledBox from '../../components/shared/boxes'
 
-import { RootState } from '../../../store'
+import { RootState } from '../../store'
 
 
-const RegisterPage: FC = () => {
+const PasswordResetPage: FC = () => {
   const errorMessage = useSelector((state: RootState) => state.message.text)
 
   const getForm = ({ isSubmitting }: { isSubmitting: boolean }) => (
@@ -39,16 +33,6 @@ const RegisterPage: FC = () => {
           <InputField name="email" type="email" />
         </FormControl>
 
-        <FormControl isRequired>
-          <FormLabel htmlFor="password">Password</FormLabel>
-          <InputField name="password" type="password" />
-        </FormControl>
-
-        <FormControl isRequired>
-          <FormLabel htmlFor="passwordrepeat">Password wiederholen</FormLabel>
-          <InputField name="passwordrepeat" type="password" />
-        </FormControl>
-
         <Button
           type="submit"
           isLoading={isSubmitting}
@@ -56,26 +40,30 @@ const RegisterPage: FC = () => {
           width="100%"
           mb="1"
         >
-          <FormattedMessage id="Register" />
+          <FormattedMessage id="Request a password reset" />
         </Button>
       </VStack>
     </Form>
   )
 
   return (
-    <TwoColumnPage>
-
+    <SplitColumns>
       <Center w="100%">
         <Heading>
-          <FormattedMessage id="Register" />
+          <FormattedMessage id="Forgot password?" />
         </Heading>
+      </Center>
+      <Center w="100%">
+        <Text align="center">
+          <FormattedMessage id="We'll send you an email with a link to reset your password." />
+        </Text>
       </Center>
 
       {errorMessage
         && (
         <StyledBox variant="errorAlert">
           <Text variant="white" fontWeight={700}>
-            <FormattedMessage id="Register not possible" />
+            <FormattedMessage id="Password reset not possible" />
           </Text>
           <Text variant="white">{errorMessage}</Text>
         </StyledBox>
@@ -84,8 +72,7 @@ const RegisterPage: FC = () => {
       <StyledBox variant="formBox">
         <Formik
           initialValues={{
-            username: `${process.env.USER_EMAIL || ''}`,
-            password: `${process.env.USER_PASSWORD || ''}`,
+            email: `${process.env.USER_EMAIL || ''}`,
           }}
           onSubmit={(values) => {
             console.log(values)
@@ -94,25 +81,9 @@ const RegisterPage: FC = () => {
           {getForm}
         </Formik>
       </StyledBox>
-
-      <Box>
-        <Center>
-          <Text>
-            <FormattedMessage id="Already registered?" />
-          </Text>
-        </Center>
-        <Center>
-          <Text>
-            <Link as={RouterLink} to="/login">
-              <FormattedMessage id="Log In" />
-            </Link>
-          </Text>
-        </Center>
-      </Box>
-
-    </TwoColumnPage>
+    </SplitColumns>
   )
 }
 
 
-export default RegisterPage
+export default PasswordResetPage
