+++ cosinnus_frontend/src/pages/register/index.tsx
@@ -1,6 +1,10 @@
 import React, { FC } from 'react'
 import { useSelector } from 'react-redux'
+
+import { Link as RouterLink } from 'react-router-dom'
+
 import {
+  Box,
   Heading,
   Button,
   VStack,
@@ -8,21 +12,23 @@ import {
   FormLabel,
   Center,
   Text,
+  Link,
 } from '@chakra-ui/react'
+
 import {
   Formik,
   Form,
 } from 'formik'
 
 import { FormattedMessage } from 'react-intl'
-import { InputField } from '../../shared/input'
-import { TwoColumnPage } from '../../shared/pages'
-import { StyledBox } from '../../shared/boxes'
+import InputField from '../../components/shared/input'
+import SplitColumns from '../../layouts/SplitColumns'
+import StyledBox from '../../components/shared/boxes'
 
-import { RootState } from '../../../store'
+import { RootState } from '../../store'
 
 
-const PasswordResetPage: FC = () => {
+const RegisterPage: FC = () => {
   const errorMessage = useSelector((state: RootState) => state.message.text)
 
   const getForm = ({ isSubmitting }: { isSubmitting: boolean }) => (
@@ -33,6 +39,16 @@ const PasswordResetPage: FC = () => {
           <InputField name="email" type="email" />
         </FormControl>
 
+        <FormControl isRequired>
+          <FormLabel htmlFor="password">Password</FormLabel>
+          <InputField name="password" type="password" />
+        </FormControl>
+
+        <FormControl isRequired>
+          <FormLabel htmlFor="passwordrepeat">Password wiederholen</FormLabel>
+          <InputField name="passwordrepeat" type="password" />
+        </FormControl>
+
         <Button
           type="submit"
           isLoading={isSubmitting}
@@ -40,31 +56,26 @@ const PasswordResetPage: FC = () => {
           width="100%"
           mb="1"
         >
-          <FormattedMessage id="Request a password reset" />
+          <FormattedMessage id="Register" />
         </Button>
       </VStack>
     </Form>
   )
 
   return (
-    <TwoColumnPage>
+    <SplitColumns>
 
       <Center w="100%">
         <Heading>
-          <FormattedMessage id="Forgot password?" />
+          <FormattedMessage id="Register" />
         </Heading>
-      </Center>
-      <Center w="100%">
-        <Text align="center">
-          <FormattedMessage id="We'll send you an email with a link to reset your password." />
-        </Text>
       </Center>
 
       {errorMessage
         && (
         <StyledBox variant="errorAlert">
           <Text variant="white" fontWeight={700}>
-            <FormattedMessage id="Password reset not possible" />
+            <FormattedMessage id="Register not possible" />
           </Text>
           <Text variant="white">{errorMessage}</Text>
         </StyledBox>
@@ -73,7 +84,8 @@ const PasswordResetPage: FC = () => {
       <StyledBox variant="formBox">
         <Formik
           initialValues={{
-            email: `${process.env.USER_EMAIL || ''}`,
+            username: `${process.env.USER_EMAIL || ''}`,
+            password: `${process.env.USER_PASSWORD || ''}`,
           }}
           onSubmit={(values) => {
             console.log(values)
@@ -82,9 +94,25 @@ const PasswordResetPage: FC = () => {
           {getForm}
         </Formik>
       </StyledBox>
-    </TwoColumnPage>
+
+      <Box>
+        <Center>
+          <Text>
+            <FormattedMessage id="Already registered?" />
+          </Text>
+        </Center>
+        <Center>
+          <Text>
+            <Link as={RouterLink} to="/login">
+              <FormattedMessage id="Log In" />
+            </Link>
+          </Text>
+        </Center>
+      </Box>
+
+    </SplitColumns>
   )
 }
 
 
-export default PasswordResetPage
+export default RegisterPage
